// src/db/database.types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export interface Database {
  graphql_public: {
    Tables: Record<never, never>;
    Views: Record<never, never>;
    Functions: {
      graphql: {
        Args: {
          operationName?: string;
          query?: string;
          variables?: Json;
          extensions?: Json;
        };
        Returns: Json;
      };
    };
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
  public: {
    Tables: {
// ...
      guides: {
        Row: {
          cover_image_url: string | null;
          created_at: string;
          creator_id: string;
          deleted_at: string | null;
          description: string;
          id: string;
          is_published: boolean;
          language: string;
          location_name: string;
          price: number;
          recommended_days: number;
          title: string;
          updated_at: string;
          version: number;
        };
        Insert: {
          cover_image_url?: string | null;
          created_at?: string;
          creator_id: string;
          deleted_at?: string | null;
          description: string;
          id?: string;
          is_published?: boolean;
          language?: string;
          location_name: string;
          price?: number;
          recommended_days?: number;
          title: string;
          updated_at?: string;
          version?: number;
        };
        Update: {
          cover_image_url?: string | null;
          created_at?: string;
          creator_id?: string;
          deleted_at?: string | null;
          description?: string;
          id?: string;
          is_published?: boolean;
          language?: string;
          location_name?: string;
          price?: number;
          recommended_days?: number;
          title?: string;
          updated_at?: string;
          version?: number;
        };
        Relationships: [
          {
            foreignKeyName: "guides_creator_id_fkey";
            columns: ["creator_id"];
            isOneToOne: false;
            referencedRelation: "creators";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
      plans: {
        Row: {
          ai_generation_cost: number | null;
          content: Json;
          created_at: string;
          deleted_at: string | null;
          generation_params: Json;
          guide_id: string;
          id: string;
          is_favorite: boolean;
          name: string;
          updated_at: string;
          user_id: string;
        };
        Insert: {
          ai_generation_cost?: number | null;
          content: Json;
          created_at?: string;
          deleted_at?: string | null;
          generation_params: Json;
          guide_id: string;
          id?: string;
          is_favorite?: boolean;
          name: string;
          updated_at?: string;
          user_id: string;
        };
        Update: {
          ai_generation_cost?: number | null;
          content?: Json;
          created_at?: string;
          deleted_at?: string | null;
          generation_params?: Json;
          guide_id?: string;
          id?: string;
          is_favorite?: boolean;
          name?: string;
          updated_at?: string;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "plans_guide_id_fkey";
            columns: ["guide_id"];
            isOneToOne: false;
            referencedRelation: "guides";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "plans_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "users";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
}